renamed bind to compose to reflect composition
